+++ src/components/stats/BudgetStats.tsx
@@ -5,18 +5,20 @@ import { GraphExpensesTimeLine } from "./GraphExpensesTimeLine";
 import { GraphDaysPerCountry } from "./GraphDaysPerCountry";
 import { BudgetModel } from "../../domain/BudgetModel";
 import Grid from "@material-ui/core/Grid";
+import { GraphDailyAverageByCountry } from "./GraphDailyAverageByCountry";
 
-interface BudgetStatsProps {
+interface BudgetStatsComponentProps {
     budget: BudgetModel;
 }
 
-export const BudgetStats: React.FC<BudgetStatsProps> = (props) => {
-    const {budget} = props; 
+export const BudgetStatsComponents: React.FC<BudgetStatsComponentProps> = (props) => {
+    const {budget} = props;
 
     return <Grid container spacing={3}>
-        {budget && <Grid item xs={12} sm={6}><GraphByCategory budget={budget} /></Grid>}
-        {budget && <Grid item xs={12} sm={6}><GraphByCountry budget={budget}/></Grid>}
-        {budget && <Grid item xs={12} sm={6}><GraphDaysPerCountry budget={budget}/></Grid>}
-        {budget && <Grid item xs={12} sm={6}><GraphExpensesTimeLine budget={budget}/></Grid>}
+        <Grid item xs={12} sm={6}><GraphByCategory budget={budget} /></Grid>
+        <Grid item xs={12} sm={6}><GraphDailyAverageByCountry budget={budget}/></Grid>
+        <Grid item xs={12} sm={6}><GraphByCountry budget={budget}/></Grid>
+        <Grid item xs={12} sm={6}><GraphDaysPerCountry budget={budget}/></Grid>
+        <Grid item xs={12} sm={12}><GraphExpensesTimeLine budget={budget}/></Grid>
     </Grid>;
 }