+++ src/components/stats/GraphDaysPerCountry.tsx
@@ -1,6 +1,7 @@
 import * as React from "react";
 import { BudgetModel } from "../../domain/BudgetModel";
-import { GraphPie } from "./Graph";
+import { getTotalDaysByCountry } from "../../domain/stats/getTotalDaysByCountry";
+import { PieChart } from "./charts/Pie";
 
 interface GraphDaysPerCountryProps {
     budget: BudgetModel
@@ -9,12 +10,18 @@ interface GraphDaysPerCountryProps {
 export const GraphDaysPerCountry: React.FC<GraphDaysPerCountryProps> = (props) => {
     const {budget} = props;
 
-    function getData () {
-        const daysByCountry = budget.totalDaysByCountry;
-        return Object
+    const data = React.useMemo(() => {
+        const daysByCountry = getTotalDaysByCountry(budget);
+        const labels: string[] = [];
+        const values: number[] = [];
+        Object
             .entries(daysByCountry)
-            .map(([country, total]) => ({x: country, y: total}));
-    }
-    
-    return <GraphPie title='Days in a country' data={getData()} />;
+            .forEach(([country, total]) => {
+                labels.push(country);
+                values.push(Math.round(total));
+            });
+        return {labels, values};
+    }, [budget]);
+
+    return <PieChart title='Days in a country' {...data} />;
 }